eliminate compiler warning introduced by my "kernel pagefault tracepoint for x86 & x86_64 patch"
Patchwork-id: 25792
O-Subject: [RHEL6 Patch] eliminate compiler warning introduced by my early
	RHEL6 "kernel pagefault tracepoint for x86 & x86_64 patch".
Bugzilla: 526032
RH-Acked-by: Rik van Riel <riel@redhat.com>

Back on Fri, 08 Jan 2010 I sent a patch to this list that adds a trace
point to the kernel pagefault path(BZ/Feature Request 526032).  This
patch adds a compiler warning because its missing a cast.

----------------------------------------------------------------------
In file included from include/trace/ftrace.h:398,
                 from include/trace/define_trace.h:69,
                 from include/trace/events/kmem.h:780,
                 from mm/page_alloc.c:57:
include/trace/events/kmem.h: In function
‘ftrace_raw_output_id_mm_kernel_pagefault’:
include/trace/events/kmem.h:755: warning: format ‘%lx’ expects type
‘long unsigned int’, but argument 5 has type ‘struct pt_regs *’
----------------------------------------------------------------------

I know its not all that important but since its one of very few warnings
I'd like to fix it now:

+++ include/trace/events/kmem.h
@@ -772,7 +772,7 @@ TRACE_EVENT(mm_kernel_pagefault,
 
 	TP_printk("task=%lx, address=%lx, regs=%lx",
 		(unsigned long)__entry->task, (unsigned long)__entry->address,
-			__entry->regs)
+			(unsigned long)__entry->regs)
 	);
 #endif /* _TRACE_KMEM_H */
